+++ src/libs/Detach.AlExtensions/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This library uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.0
+
+### Added
+
+- Added `BufferFormatUtils.GetBufferFormat` method.
+
 ## 0.2.0
 
 ### Changed